Update patch set 3
Patch Set 3:

> > > > I think this is not the right place to do this kind of mapping.
> > > > Should probably be in build/soong/cc/config/arm_device.go
> > > > There's similar logic there already (e.g. for "kyro"): https://android.googlesource.com/platform/build/soong/+/master/cc/config/arm_device.go#71
> > > 
> > > In this case, GNU as is invoked directly by Clang, due to "-no-integrated-as" flag. Soong cannot override the flags passed by Clang.
> > 
> > The mapping in build/soong/cc/config/arm_device.go is for Cflags, so it'll affect what clang gets as argument, which in turn would affect what clang passes to as.
> > 
> > Maybe I'm missing something, and this variable works in some other way.
> 
> That's how it works.  But we want to pass mcpu=cortex-a55 to Clang so when the code gets tuned for a55.  But we also want to avoid breaking assembly when invoking the Gnu Assembler.  In the past, we had the Clang driver pass the similar-but-old CPU to Gnu as (https://reviews.llvm.org/D40476).  But it's too late now to update Clang and we instead want to use a wrapper for 'as' to do the rewrite.

I don't know the difference between cortex-a55 and cortex-a53, but if it has new instructions, and clang generates these instructions, and passes them to the assembler with -mcpu=cortex-a53 flag, how is the assembler going to assemble them? If the assembler doesn't support cortex-a55, it is going to error out on the input anyway. Maybe I'm missing something again.

If we really want to pass -mcpu=cortex-a55 to clang, and make the assembler treat it the same as -mcpu=cortex-a53, the right place for the change would still not be here. The mapping should be in the binutils source repo (preferably a backport of an upstream commit that does the same). As it stands right now, this wrapper will get removed the next time binutils prebuilts are updated.

Patch-set: 3